pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datacasketshred"
version = "0.1.0"
description = "Archive files into password-protected 7z and securely delete originals."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "DataCasketShred Maintainer" }]
dependencies = [
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
